+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -13,6 +13,7 @@ import jakarta.persistence.Transient;
 
 import java.math.BigDecimal;
 import java.util.UUID;
+import kitchenpos.menu.domain.Menu;
 
 @Table(name = "order_line_item")
 @Entity
