+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -1,10 +1,13 @@
-package kitchenpos.application;
+package kitchenpos.menu.application;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.application.FakePurgomalumClient;
+import kitchenpos.application.InMemoryProductRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.infra.InMemoryMenuRepository;
 import kitchenpos.menugroup.infra.InMemoryMenuGroupRepository;
 import kitchenpos.menugroup.domain.MenuGroupRepository;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.MenuRepository;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.ProductRepository;
 import kitchenpos.infra.PurgomalumClient;
